culturemeshFFB/requirements.txt
alabaster==0.7.11
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
blinker==1.4
certifi==2018.8.13
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.2
docutils==0.14
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
flup-py3==1.0.3
gunicorn==19.7.1
idna==2.6
imagesize==1.0.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
nose==1.3.7
packaging==17.1
passlib==1.7.1
pbr==5.1.1
pluggy==0.8.0
py==1.7.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==4.1.0
pytest-cov==2.6.1
pytest-mock==1.10.0
python-dateutil==2.7.3
pytz==2018.3
qpp-git-secrets==1.1.1
requests==2.20.1
six==1.11.0
snowballstemmer==1.2.1
speaklater==1.3
Sphinx==1.7.6
sphinx-rtd-theme==0.4.1
sphinxcontrib-websupport==1.1.0
unidiff==0.5.5
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.1
